Moves Footer from partials to components #476
