+++ node/magnet/src/components/Footer.js
