+++ node/magnet/src/components/Footer.soy
